+++ test/Armada.Test.Unit/Program.cs
@@ -257,6 +257,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new IncidentServiceTests());
             runner.AddSuite(new ObjectiveServiceTests());
             runner.AddSuite(new KeyedAsyncLockTests());
+            runner.AddSuite(new ObjectiveTerminalBacklogStateTests());
             runner.AddSuite(new ObjectiveRefinementCoordinatorTests());
             runner.AddSuite(new ObjectiveDependencyAnalyzerTests());
             runner.AddSuite(new ObjectiveBriefRendererTests());